+++ lib/ts/recipe/emailverification/emaildelivery/services/index.ts
@@ -14,4 +14,4 @@
  */
 
 import SMTP from "./smtp";
-export let STMPService = SMTP;
+export let SMTPService = SMTP;