+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [11.1.1] - 2022-08-11
+
 ### Bug fix:
 
 -   Adds missing `force` param from `createUserIdMapping` and `deleteUserIdMapping` functions